requirements.txt
google-genai
python-dotenv
chromadb
requests
langchain
PyPDF2
